Create tensorflow module
