+++ vehicle-python/src/vehicle_lang/compile/python.py
@@ -5,8 +5,10 @@ from pathlib import Path
 from typing import (
     Any,
     Dict,
+    Iterable,
     Iterator,
     List,
+    Optional,
     Sequence,
     SupportsFloat,
     SupportsInt,
@@ -17,13 +19,45 @@ from typing import (
 from typing_extensions import Self, final, override
 
 from .. import ast as vcl
-from ..typing import AnyOptimisers, Optimiser
+from ..ast import (
+    AST,
+    Binder,
+    BuiltinFunction,
+    Declaration,
+    Expression,
+    Program,
+    Provenance,
+)
+from ..typing import (
+    AnyOptimisers,
+    DeclarationName,
+    DifferentiableLogic,
+    Explicit,
+    Target,
+)
 from ._ast_compat import arguments as py_arguments
 from ._ast_compat import dump as py_ast_dump
 from ._ast_compat import unparse as py_ast_unparse
 from .abc import ABCTranslation, AnyBuiltins
 from .abcboolasbool import ABCBoolAsBoolBuiltins
-from .error import VehicleOptimiseTypeError
+from .error import VehicleOptimiseTypeError, VehiclePropertyNotFound
+
+__all__: List[str] = [
+    # Abstract Syntax Tree
+    "AST",
+    "Binder",
+    "BuiltinFunction",
+    "Declaration",
+    "Expression",
+    "Program",
+    "Provenance",
+    # Translation to Python
+    "PythonBuiltins",
+    "PythonTranslation",
+    # High-level functions
+    "compile",
+    "load_loss_function",
+]
 
 ################################################################################
 ### Implementation of Vehicle builtins in Python
@@ -482,3 +516,46 @@ def py_partial_app(
         *arguments,
         provenance=provenance,
     )
+
+
+def load(
+    path: Union[str, Path],
+    *,
+    declarations: Iterable[DeclarationName] = (),
+    target: Target = Explicit.Explicit,
+    translation: Optional[PythonTranslation] = None,
+) -> Dict[str, Any]:
+    if translation is None:
+        translation = PythonTranslation(builtins=PythonBuiltins(optimisers={}))
+    return translation.compile(
+        vcl.load(path, declarations=declarations, target=target), path=path
+    )
+
+
+def load_loss_function(
+    path: Union[str, Path],
+    property_name: DeclarationName,
+    *,
+    target: DifferentiableLogic = DifferentiableLogic.Vehicle,
+    optimisers: AnyOptimisers = {},
+) -> Any:
+    """
+    Load a loss function from a property in a Vehicle specification.
+
+    :param path: The path to the Vehicle specification file.
+    :param property_name: The name of the Vehicle property to load.
+    :param target: The differentiable logic to use for interpreting the Vehicle property as a loss function, defaults to the Vehicle logic.
+    :param samplers: A map from quantified variable names to samplers for their values. See `Sampler` for more details.
+    :return: A function that takes the required external resources in the specification as keyword arguments and returns the loss corresponding to the property.
+    """
+    translation = PythonTranslation(builtins=PythonBuiltins(optimisers=optimisers))
+    declarations = load(
+        path,
+        declarations=(property_name,),
+        target=target,
+        translation=translation,
+    )
+    if property_name in declarations:
+        return declarations[property_name]
+    else:
+        raise VehiclePropertyNotFound(property_name)
