+++ vehicle-python/src/vehicle_lang/tensorflow.py
@@ -1,9 +1,8 @@
 from typing import List
 
-from ._version import VERSION as VERSION
 from .compile.error import VehicleBuiltinUnsupported as VehicleBuiltinUnsupported
 from .compile.error import VehiclePropertyNotFound as VehiclePropertyNotFound
-from .compile.python import load_loss_function as load_loss_function
+from .compile.tensorflow import load_loss_function as load_loss_function
 from .error import VehicleError as VehicleError
 from .error import VehicleInternalError as VehicleInternalError
 from .session.error import VehicleSessionClosed as VehicleSessionClosed
@@ -14,15 +13,10 @@ from .typing import DeclarationName as DeclarationName
 from .typing import DifferentiableLogic as DifferentiableLogic
 from .typing import Optimiser as Optimiser
 from .typing import QuantifiedVariableName as QuantifiedVariableName
-from .typing import Verifier as Verifier
-from .verify import verify as verify
 
 __all__: List[str] = [
-    "VERSION",
     # Compile
     "load_loss_function",
-    # Verify
-    "verify",
     # Error types
     "VehicleError",
     "VehicleSessionClosed",
@@ -37,5 +31,4 @@ __all__: List[str] = [
     "AnyOptimiser",
     "AnyOptimisers",
     "DifferentiableLogic",
-    "Verifier",
 ]
