+++ vehicle-python/tests/test_lossdl2_exec.py
@@ -1,10 +1,11 @@
 from functools import reduce
 from pathlib import Path
-from typing import Any, Callable, Dict, Iterator, Union
+from typing import Any, Callable, Dict, Union
 
 import pytest
+
 import vehicle_lang as vcl
-import vehicle_lang.compile as vclc
+import vehicle_lang.compile.python as vcl2py
 
 
 def network_validate_output(output: Dict[str, Any]) -> None:
@@ -118,10 +119,10 @@ def test_loss_function_exec(
 ) -> None:
     print(f"Exec {specification_filename}")
     specification_path = Path(__file__).parent / "data" / specification_filename
-    actual_declarations = vclc.compile(
+    actual_declarations = vcl2py.load(
         specification_path,
         target=vcl.DifferentiableLogic.DL2,
-        translation=vclc.PythonTranslation.from_optimisers(optimisers),
+        translation=vcl2py.PythonTranslation.from_optimisers(optimisers),
     )
     if isinstance(validate_output, dict):
         for key in validate_output.keys():
